apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
bcrypt==4.2.0
Pillow==10.4.0
python-multipart==0.0.9
